fix(31-07): B-GAP-2 pointer-truncations (2 FUNCTIONAL x64 fixes + shared header)
FUNCTIONAL x64 corruptions:
- PathSearch/PathNode (UAF): m_marks[4] int -> intptr_t end-to-end (field +
  getMark/setMark accessors + all PathSearch.cpp casts). Mark slot 3 round-trips
  a live PathSearchNode*; an int slot truncated it on x64 -> use-after-free.
  Marks are runtime-only ('not persisted') so no serialization width concern.
- StatusWindow (this): SetWindowLong(GWL_USERDATA, reinterpret_cast<LONG>(this))
  + the read-back -> SetWindowLongPtr/GetWindowLongPtr + GWLP_USERDATA + LONG_PTR.
  A 32-bit userdata slot truncated  on x64 -> corrupt message dispatch.

Width-correct:
- VoidBindSecond.h (SHARED TEMPLATE HEADER, C4312): integral->pointer bind now
  routes through uintptr_t via an if-constexpr convertBindArgument<> helper
  (the EditableAnimationState  int->T* instantiation).
  Template-internal change only (no layout); plan 31-06 Task 2 full 5-target
  build recompiles all consumers (ABI-cascade noted).
- LfgDataTable: matchLevel void* level-bounds unpack unsigned long -> uintptr_t
  (unsigned long is 32-bit LLP64; small-value compare identical both bitnesses).
- AlterScheduler: %x reinterpret_cast<unsigned int>(object) -> %p + void const*.

All touched TUs 0 C4311/C4312 x64; Archive.h D-07 exclusion untouched.
DEF-31-07: 2 unrelated std::distance->int C4244 (D-07/N2 class) logged, NOT fixed.

+++ src/engine/shared/library/sharedGame/src/shared/core/LfgDataTable.cpp
@@ -105,7 +105,12 @@ namespace LfgDataTableNamespace
 
 	bool matchLevel(LfgCharacterData const & lfgCharacterData, void const * lowLevel, void const * highLevel, void const *, void const *, void const *)
 	{
-		return ((static_cast<unsigned long>(lfgCharacterData.level) >= reinterpret_cast<unsigned long>(lowLevel)) && (static_cast<unsigned long>(lfgCharacterData.level) <= reinterpret_cast<unsigned long>(highLevel)));
+		// BITS-02 B-GAP-2: the level bounds are integers PACKED into the void*
+		// search-attribute slots. Unpack them at full pointer width (uintptr_t);
+		// `unsigned long` is 32-bit under LLP64 and truncated the 64-bit pointer
+		// slot on x64 (C4311). Character levels are small, so the value compare
+		// is identical on both bitnesses.
+		return ((static_cast<uintptr_t>(lfgCharacterData.level) >= reinterpret_cast<uintptr_t>(lowLevel)) && (static_cast<uintptr_t>(lfgCharacterData.level) <= reinterpret_cast<uintptr_t>(highLevel)));
 	}
 
 	bool matchFriend(LfgCharacterData const & lfgCharacterData, void const * pFriendList, void const *, void const *, void const *, void const *)
